Split spec models into their own files
