+++ spec/support/models/vegetable.rb
@@ -1,0 +1,14 @@
+require 'support/active_record_schema'
+
+ar_schema.create_table :vegetables do |t|
+  t.string :name
+end
+
+class Vegetable < ActiveRecord::Base
+  include MeiliSearch::Rails
+
+  meilisearch raise_on_failure: false, index_uid: safe_index_uid('Fruit') do
+    attribute :name
+  end
+end
+
